add file lock
